more cleaning

+++ src/natives/java/lang/ClassLoader.c
@@ -9,7 +9,7 @@ DECLARE_NATIVE("java/lang", ClassLoader, findLoadedClass0,
   assert(argc == 1);
   heap_string read = read_string_to_utf8(args[0].handle->obj);
   // Replace . with /
-  for (size_t i = 0; i < read.len; ++i)
+  for (int i = 0; i < read.len; ++i)
     if (read.chars[i] == '.')
       read.chars[i] = '/';
   bjvm_classdesc *cd = bootstrap_class_create(
